+++ requirements.txt
@@ -10,29 +10,7 @@ idna==2.9
 itsdangerous==1.1.0
 Jinja2==2.11.1
 MarkupSafe==1.1.1
-pkg-resources==0.0.0
-psycopg2==2.8.4
-python-dateutil==2.8.1
-requests==2.23.0
-six==1.14.0
-SQLAlchemy==1.3.13
-text-unidecode==1.3
-urllib3==1.25.8
-Werkzeug==1.0.0
-blinker==1.4
-certifi==2019.11.28
-chardet==3.0.4
-Click==7.0
-Faker==4.0.1
-Flask==1.1.1
-Flask-DebugToolbar==0.11.0
-Flask-SQLAlchemy==2.4.1
-idna==2.9
-itsdangerous==1.1.0
-Jinja2==2.11.1
-MarkupSafe==1.1.1
-pkg-resources==0.0.0
-psycopg2==2.8.4
+psycopg2-binary==2.8.4
 python-dateutil==2.8.1
 requests==2.23.0
 six==1.14.0
